Fix deployment migrations

+++ database/migrations/2025_04_26_151128_create_items_table.php
@@ -8,22 +8,24 @@ class CreateItemsTable extends Migration
 {
     public function up()
     {
-        Schema::create('items', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->foreignId('category_id')->constrained()->onDelete('cascade');
-            $table->foreignId('subcategory_id')->constrained()->onDelete('cascade');
-            $table->foreignId('group_id')->constrained()->onDelete('cascade')->nullable();
-            $table->string('model')->nullable();
-            $table->enum('status', ['Available', 'Rented', 'Damaged', 'Lost'])->default('Available');
-            $table->string('serial_number')->nullable();
-            $table->enum('unit', ['Kg', 'Cartons', 'PC', 'L','M','Sqm'])->default('PC');
-            $table->integer('quantity')->default(0);
-            $table->string('flight_case_number')->nullable();
-            $table->text('remarks')->nullable();
-            $table->string('image')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('items')) {
+            Schema::create('items', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->foreignId('category_id')->constrained()->onDelete('cascade');
+                $table->foreignId('subcategory_id')->constrained()->onDelete('cascade');
+                $table->foreignId('group_id')->nullable()->constrained()->onDelete('cascade');
+                $table->string('model')->nullable();
+                $table->enum('status', ['Available', 'Rented', 'Damaged', 'Lost'])->default('Available');
+                $table->string('serial_number')->nullable();
+                $table->enum('unit', ['Kg', 'Cartons', 'PC', 'L', 'M', 'Sqm'])->default('PC');
+                $table->integer('quantity')->default(0);
+                $table->string('flight_case_number')->nullable();
+                $table->text('remarks')->nullable();
+                $table->string('image')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
